Refactor GameLoad, Able to Slide Top and Bottom
