--- app/src/main/java/com/delacrixmorgan/kingscup/common/FragmentListener.kt
@@ -1,13 +1,0 @@
-package com.delacrixmorgan.kingscup.common
-
-/**
- * FragmentListener
- * kingscup-android
- *
- * Created by Delacrix Morgan on 25/03/2018.
- * Copyright (c) 2018 licensed under a Creative Commons Attribution-ShareAlike 4.0 International License.
- */
-
-interface FragmentListener {
-    fun onBackPressed()
-}
