+++ app/src/main/java/com/delacrixmorgan/kingscup/game/GameLoadFragment.kt
@@ -2,49 +2,26 @@ package com.delacrixmorgan.kingscup.game
 
 import android.os.Bundle
 import android.os.Handler
-import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.navigation.Navigation
 import com.delacrixmorgan.kingscup.R
 import com.delacrixmorgan.kingscup.common.GameEngine
 import com.delacrixmorgan.kingscup.common.SoundEngine
 import com.delacrixmorgan.kingscup.common.SoundType
-import com.delacrixmorgan.kingscup.common.showFragmentSliding
-import com.delacrixmorgan.kingscup.model.LoadType
 import kotlinx.android.synthetic.main.fragment_game_load.*
 
 /**
  * GameLoadFragment
  * kingscup-android
  *
- * Created by Delacrix Morgan on 25/03/2018.
- * Copyright (c) 2018 licensed under a Creative Commons Attribution-ShareAlike 4.0 International License.
+ * Created by Delacrix Morgan on 08/05/2019.
+ * Copyright (c) 2019 licensed under a Creative Commons Attribution-ShareAlike 4.0 International License.
  */
 
 class GameLoadFragment : Fragment() {
-
-    companion object {
-        private const val GAME_LOAD_TYPE = "GameLoadFragment.Type"
-
-        fun newInstance(loadType: LoadType): GameLoadFragment {
-            val fragment = GameLoadFragment()
-            val args = Bundle()
-
-            args.putSerializable(GAME_LOAD_TYPE, loadType)
-            fragment.arguments = args
-
-            return fragment
-        }
-    }
-
-    private lateinit var loadType: LoadType
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        this.loadType = arguments?.getSerializable(GAME_LOAD_TYPE) as LoadType
-    }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         return inflater.inflate(R.layout.fragment_game_load, container, false)
@@ -52,16 +29,22 @@ class GameLoadFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val context = this.context ?: return
+        val context = view.context
 
-        this.loadingTextView.text = this.loadType.getLocalisedText(context)
-        GameEngine.newInstance(context)
-        SoundEngine.getInstance().playSound(context, SoundType.KING)
+        this.arguments?.let {
+            val gameType = GameLoadFragmentArgs.fromBundle(it).gameType
+            this.loadingTextView.text = gameType.getLocalisedText(context)
+
+            GameEngine.newInstance(context)
+            SoundEngine.getInstance().playSound(context, SoundType.KING)
+        }
 
         Handler().postDelayed({
             run {
-                this.activity?.supportFragmentManager?.popBackStack()
-                this.context?.showFragmentSliding(GameBoardFragment.newInstance(), Gravity.TOP)
+                if (this.isVisible) {
+                    val action = GameLoadFragmentDirections.actionGameLoadFragmentToGameBoardFragment()
+                    Navigation.findNavController(view).navigate(action)
+                }
             }
         }, 2000)
     }
