+++ app/src/main/java/com/delacrixmorgan/kingscup/model/Card.kt
@@ -1,6 +1,5 @@
 package com.delacrixmorgan.kingscup.model
 
-import android.annotation.SuppressLint
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
@@ -12,6 +11,5 @@ import kotlinx.android.parcel.Parcelize
  * Copyright (c) 2018 licensed under a Creative Commons Attribution-ShareAlike 4.0 International License.
  */
 
-@SuppressLint("ParcelCreator")
 @Parcelize
 class Card(val suit: String, val rank: String, val header: String, val body: String) : Parcelable
