+++ app/src/main/res/navigation/nav_graph.xml
@@ -39,7 +39,11 @@
             app:popExitAnim="@anim/slide_out_right" />
         <action
             android:id="@+id/action_menuFragment_to_gameLoadFragment"
-            app:destination="@id/gameLoadFragment" />
+            app:destination="@id/gameLoadFragment"
+            app:enterAnim="@anim/slide_in_top"
+            app:exitAnim="@anim/slide_out_bottom"
+            app:popEnterAnim="@anim/slide_in_bottom"
+            app:popExitAnim="@anim/slide_out_top" />
     </fragment>
     <fragment
         android:id="@+id/menuRateFragment"
@@ -66,8 +70,11 @@
         <action
             android:id="@+id/action_gameLoadFragment_to_gameBoardFragment"
             app:destination="@id/gameBoardFragment"
-            app:popUpTo="@+id/gameBoardFragment"
+            app:popUpTo="@+id/gameLoadFragment"
             app:popUpToInclusive="true" />
+        <argument
+            android:name="gameType"
+            app:argType="com.delacrixmorgan.kingscup.model.GameType" />
     </fragment>
     <fragment
         android:id="@+id/gameBoardFragment"
@@ -80,6 +87,11 @@
         <action
             android:id="@+id/action_gameBoardFragment_to_menuFragment2"
             app:destination="@id/menuFragment" />
+        <action
+            android:id="@+id/action_gameBoardFragment_to_gameLoadFragment"
+            app:destination="@id/gameLoadFragment"
+            app:popUpTo="@+id/gameBoardFragment"
+            app:popUpToInclusive="true" />
     </fragment>
     <fragment
         android:id="@+id/gameCardFragment"
